rendering_tools/pyproject.toml
[project]
name = "rendering-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pycairo>=1.26.1",
    "pytest>=8.3.2",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]